refactor: move get functions to separate module, remove eval
